+++ rust-plugins/worker/options.d.ts
@@ -1,3 +1,5 @@
+import type { UserConfig } from "@farmfe/core"
 export interface IPluginOptions {
-
+  isBuild?: boolean
+  compilerConfig?: UserConfig
 }