+++ rust-plugins/worker/playground/farm.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from "@farmfe/core";
 import react from '@farmfe/plugin-react';
 import farmPlugin from '@farmfe/plugin-worker';
-
 export default defineConfig({
   compilation: {
     input: {