+++ rust-plugins/virtual/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/plugin-virtual
 
+## 0.0.7
+
+### Patch Changes
+
+- feat(rust-plugin): optimization plugin
+
 ## 0.0.6
 
 ### Patch Changes