+++ rust-plugins/wasm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/plugin-wasm
 
+## 0.0.2
+
+### Patch Changes
+
+- feat(rust-plugin): optimization plugin
+
 ## 0.0.1
 
 ### Patch Changes